refactor(store): Moves store into src folder
Fixes #274
