+++ src/extra/context-menu/index.js
@@ -1,4 +1,4 @@
-import store from '@/../store';
+import store from '@/store';
 import { Menu } from 'nwjs-menu-browser';
 // import '@/../node_modules/nwjs-menu-browser/nwjs-menu-browser.css';
 import contextMenuStore from './store';
