+++ src/extra/midi-assignment/index.js
@@ -1,4 +1,4 @@
-import store from '@/../store';
+import store from '@/store';
 import { modV } from 'modv';
 import { MenuItem } from 'nwjs-menu-browser';
 import midiAssignmentStore from './store';
