+++ src/modv/get-largest-window.js
@@ -1,4 +1,4 @@
-import store from '@/../store/';
+import store from '@/store/';
 
 /**
  * Returns window area
