+++ src/store/modules/control-panels.js
