+++ src/store/modules/palettes.js
