+++ src/store/modules/plugins/index.js
