+++ src/store/modules/plugins/statusbar.js
