+++ src/store/modules/size.js
@@ -1,4 +1,4 @@
-import store from '@/../store';
+import store from '@/store';
 import { modV } from '@/modv';
 import Vue from 'vue';
