+++ src/store/modules/user.js
@@ -1,4 +1,4 @@
-import store from '@/../store';
+import store from '@/store';
 import { modV } from '@/modv';
 
 const state = {
